cleanup StoppingState, moved it to the separate mod
